+++ package.json
@@ -89,10 +89,10 @@
     "babel-plugin-tester": "^8.0.1",
     "detox": "^16.0.0",
     "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.2.0",
+    "eslint-config-prettier": "^6.10.1",
     "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.0",
-    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react-hooks": "^2.5.0",
     "husky": "^3.0.8",
     "jest": "^25.1.0",
